Fix lpj_guess.yaml: move further_reading inside lpj_guess block
The further_reading key was at the top level, causing it to be
treated as a model name. This led to TypeError when check_model_lresume
tried to access config['further_reading']['lresume'] since further_reading
is a list, not a dict.

Moved further_reading inside the lpj_guess: block and fixed the path
format to match other components.